REF Rename math.js in utils.js with findDuplicates in it
