+++ src/utils.js
@@ -1,0 +1,18 @@
+const findDuplicates = (array) => {
+  if ( array.length < 2 ) {
+    return [];
+  }
+
+  const duplicates = [];
+  for (let counter = 0 ; counter < array.length - 1 ; counter++) {
+    const lastIndex = array.lastIndexOf(array[counter]);
+    if (counter !== lastIndex) {
+      duplicates.push(array[counter]);
+    }
+  }
+  return duplicates;
+};
+
+const random = (max) => Math.floor(Math.random() * max);
+
+export { findDuplicates, random };
