--- src/utils/math.js
@@ -1,3 +1,0 @@
-const random = (max) => Math.floor(Math.random() * max);
-
-export { random };
